backend/requirement.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
pypdf
python-dotenv
langchain_text_splitters